+++ packages/core/vite.config.ts
@@ -1,0 +1,9 @@
+/* Copyright 2021, Milkdown by Mirone. */
+import path from 'path';
+import { defineConfig } from 'vite';
+
+import { viteBuild } from '../../vite.config.common';
+
+export default defineConfig({
+    build: viteBuild('core'),
+});