+++ packages/design-system/src/types.ts
@@ -38,10 +38,13 @@ export type Icon =
     | 'unchecked'
     | 'checked'
     | 'undo'
-    | 'redo';
+    | 'redo'
+    | 'liftList'
+    | 'sinkList';
 
 export type Slots = {
     icon: (id: Icon, config?: Record<string, string | number | boolean>) => HTMLElement;
+    label: (id: Icon, config?: Record<string, string | number | boolean>) => string;
 };
 
 export type MixinFactory = {