+++ packages/plugin-clipboard/vite.config.ts
@@ -1,0 +1,9 @@
+/* Copyright 2021, Milkdown by Mirone. */
+
+import { defineConfig } from 'vite';
+
+import { viteBuild } from '../../vite.config.common';
+
+export default defineConfig({
+    build: viteBuild('plugin-clipboard'),
+});