+++ packages/plugin-menu/src/menubar.ts
@@ -5,33 +5,36 @@ import { Utils } from '@milkdown/utils';
 
 export const menubar = (utils: Utils, view: EditorView) => {
     const editorWrapperStyle = utils.getStyle((themeTool) => {
-        return css`
-            ${themeTool.mixin.scrollbar('y')};
-        `;
+        return themeTool.mixin.scrollbar('y');
     });
     const menuStyle = utils.getStyle((themeTool) => {
-        return css`
+        const border = themeTool.mixin.border();
+        const scrollbar = themeTool.mixin.scrollbar('x');
+        const style = css`
             box-sizing: border-box;
             width: 100%;
             display: flex;
             flex-wrap: nowrap;
             overflow-x: auto;
-            ${themeTool.mixin.border()};
-            ${themeTool.mixin.scrollbar('x')};
+            ${border};
+            ${scrollbar};
+
             -webkit-overflow-scrolling: auto;
 
             .disabled {
                 display: none;
             }
         `;
+        return style;
     });
 
     const menuWrapper = document.createElement('div');
     const menu = document.createElement('div');
+    menu.classList.add('milkdown-menu');
     menuWrapper.appendChild(menu);
     menuWrapper.classList.add('milkdown-menu-wrapper');
     if (menuStyle) {
-        menu.classList.add(menuStyle);
+        menuStyle.split(' ').forEach((x) => menu.classList.add(x));
     }
 
     const editorDom = view.dom;