+++ packages/vue/src/Editor.tsx
@@ -1,14 +1,16 @@
 /* Copyright 2021, Milkdown by Mirone. */
-import { Editor, editorViewCtx, rootCtx } from '@milkdown/core';
+import { Ctx, Editor, editorViewCtx, rootCtx } from '@milkdown/core';
 import { ViewFactory } from '@milkdown/prose';
 import {
+    ComponentInternalInstance,
     DefineComponent,
     defineComponent,
-    Fragment,
+    getCurrentInstance,
     h,
     inject,
     InjectionKey,
     markRaw,
+    onBeforeMount,
     onMounted,
     onUnmounted,
     provide,
@@ -17,16 +19,19 @@ import {
     shallowReactive,
 } from 'vue';
 
-import { PortalPair, Portals } from './Portals';
+import { AnyVueComponent } from './utils';
 import { createVueView } from './VueNodeView';
 
-const rendererKey: InjectionKey<(component: DefineComponent) => ViewFactory> = Symbol();
+const rendererKey: InjectionKey<(component: DefineComponent) => (ctx: Ctx) => ViewFactory> = Symbol();
 
-type GetEditor = (container: HTMLDivElement, renderVue: (Component: DefineComponent) => ViewFactory) => Editor;
+type GetEditor = (
+    container: HTMLDivElement,
+    renderVue: (Component: AnyVueComponent) => (ctx: Ctx) => ViewFactory,
+) => Editor;
 
 const useGetEditor = (getEditor: GetEditor) => {
     const divRef = ref<HTMLDivElement | null>(null);
-    const renderVue = inject<(Component: DefineComponent) => ViewFactory>(rendererKey, () => {
+    const renderVue = inject<(Component: DefineComponent) => (ctx: Ctx) => ViewFactory>(rendererKey, () => {
         throw new Error();
     });
     const editorRef = markRaw<{ editor?: Editor }>({});
@@ -52,39 +57,69 @@ const useGetEditor = (getEditor: GetEditor) => {
     return { divRef, editorRef };
 };
 
-export const EditorComponent = defineComponent((props: { editor: GetEditor; editorRef?: Ref<EditorRef> }) => {
-    const refs = useGetEditor(props.editor);
-    if (props.editorRef) {
-        props.editorRef.value = {
-            get: () => refs.editorRef.editor,
-            dom: () => refs.divRef.value,
-        };
-    }
+export const EditorComponent = defineComponent<{ editor: GetEditor; editorRef?: Ref<EditorRef> }>({
+    name: 'milkdown-dom-root',
+    setup: (props, { slots }) => {
+        const refs = useGetEditor(props.editor);
+        if (props.editorRef) {
+            props.editorRef.value = {
+                get: () => refs.editorRef.editor,
+                dom: () => refs.divRef.value,
+            };
+        }
 
-    return () => <div ref={refs.divRef} />;
+        return () => <div ref={refs.divRef}>{slots.default?.()}</div>;
+    },
 });
 EditorComponent.props = ['editor', 'editorRef'];
 
 export type EditorRef = { get: () => Editor | undefined; dom: () => HTMLDivElement | null };
 
-export const VueEditor = defineComponent((props: { editor: GetEditor; editorRef?: Ref<EditorRef> }) => {
-    const portals = shallowReactive<PortalPair[]>([]);
-    const addPortal = markRaw((component: DefineComponent, key: string) => {
-        portals.push([key, component]);
-    });
-    const removePortalByKey = markRaw((key: string) => {
-        const index = portals.findIndex((p) => p[0] === key);
-        portals.splice(index, 1);
-    });
-    const renderVue = createVueView(addPortal, removePortalByKey);
-    provide(rendererKey, renderVue);
+const rootInstance: {
+    instance: null | ComponentInternalInstance;
+} = {
+    instance: null,
+};
+export const getRootInstance = () => {
+    return rootInstance.instance;
+};
 
-    return () => (
-        <>
-            <Portals portals={portals} />
-            <EditorComponent editorRef={props.editorRef} editor={props.editor} />
-        </>
-    );
+type PortalPair = [key: string, component: DefineComponent];
+export const VueEditor = defineComponent<{ editor: GetEditor; editorRef?: Ref<EditorRef> }>({
+    name: 'milkdown-vue-root',
+    setup: (props) => {
+        const portals = shallowReactive<PortalPair[]>([]);
+
+        const instance = getCurrentInstance();
+
+        onBeforeMount(() => {
+            rootInstance.instance = (instance as ComponentInternalInstance & { ctx: { _: ComponentInternalInstance } })
+                .ctx._ as ComponentInternalInstance;
+        });
+
+        onUnmounted(() => {
+            rootInstance.instance = null;
+        });
+
+        const addPortal = markRaw((component: DefineComponent, key: string) => {
+            portals.push([key, component]);
+        });
+        const removePortalByKey = markRaw((key: string) => {
+            const index = portals.findIndex((p) => p[0] === key);
+            portals.splice(index, 1);
+        });
+        const renderVue = createVueView(addPortal, removePortalByKey);
+        provide(rendererKey, renderVue);
+
+        return () => {
+            const portalElements = portals.map(([id, P]) => <P key={id} />);
+            return (
+                <EditorComponent editorRef={props.editorRef} editor={props.editor}>
+                    {portalElements}
+                </EditorComponent>
+            );
+        };
+    },
 });
 VueEditor.props = ['editor', 'editorRef'];