+++ packages/vue/src/VueNodeView.tsx
@@ -1,25 +1,28 @@
 /* Copyright 2021, Milkdown by Mirone. */
+import { Ctx } from '@milkdown/core';
 import type { Decoration, EditorView, NodeView, ViewFactory } from '@milkdown/prose';
 import { Mark, Node } from '@milkdown/prose';
 import { customAlphabet } from 'nanoid';
-import { DefineComponent, defineComponent, h, Teleport } from 'vue';
+import { DefineComponent, defineComponent, h, markRaw, Teleport } from 'vue';
 
+import { getRootInstance } from '.';
 import { Content, VueNodeContainer } from './VueNode';
 
 const nanoid = customAlphabet('abcedfghicklmn', 10);
 
 export const createVueView =
     (addPortal: (portal: DefineComponent, key: string) => void, removePortalByKey: (key: string) => void) =>
-    (component: DefineComponent): ViewFactory =>
+    (component: DefineComponent): ((ctx: Ctx) => ViewFactory) =>
+    (ctx) =>
     (node, view, getPos, decorations) =>
-        new VueNodeView(component, addPortal, removePortalByKey, node, view, getPos, decorations);
+        new VueNodeView(ctx, component, addPortal, removePortalByKey, node, view, getPos, decorations);
 
 export class VueNodeView implements NodeView {
-    dom: HTMLElement | undefined;
-    contentDOM: HTMLElement | undefined;
+    teleportDOM: HTMLElement;
     key: string;
 
     constructor(
+        private ctx: Ctx,
         private component: DefineComponent,
         private addPortal: (portal: DefineComponent, key: string) => void,
         private removePortalByKey: (key: string) => void,
@@ -29,56 +32,94 @@ export class VueNodeView implements NodeView {
         private decorations: Decoration[],
     ) {
         this.key = nanoid();
-        const dom = document.createElement(node instanceof Mark ? 'span' : 'div');
-        dom.classList.add('dom-wrapper');
-
-        const contentDOM =
-            node instanceof Node && node.isLeaf
-                ? undefined
-                : document.createElement(node instanceof Mark ? 'span' : 'div');
-        if (contentDOM) {
-            contentDOM.classList.add('content-dom');
-            dom.appendChild(contentDOM);
-        }
-        this.dom = dom;
-        this.contentDOM = contentDOM;
+        this.teleportDOM = document.createElement(node instanceof Mark ? 'span' : 'div');
         this.renderPortal();
     }
 
+    get dom() {
+        return this.teleportDOM.firstElementChild || this.teleportDOM;
+    }
+
+    get contentDOM() {
+        if (this.node instanceof Node && this.node.isLeaf) {
+            return null;
+        }
+
+        return this.teleportDOM.querySelector('[data-view-content]') || this.dom;
+    }
+
     renderPortal() {
-        if (!this.dom) return;
+        if (!this.teleportDOM) return;
 
         const CustomComponent = this.component;
-        const Portal = defineComponent(() => {
-            return () => (
-                <Teleport to={this.dom}>
-                    <VueNodeContainer
-                        key={this.key}
-                        node={this.node}
-                        view={this.view}
-                        getPos={this.getPos}
-                        decorations={this.decorations}
-                    >
-                        <CustomComponent>
-                            <Content dom={this.contentDOM} />
-                        </CustomComponent>
-                    </VueNodeContainer>
-                </Teleport>
-            );
+        const Portal = defineComponent({
+            name: 'milkdown-portal',
+            setup: () => {
+                return () => (
+                    <Teleport key={this.key} to={this.teleportDOM}>
+                        <VueNodeContainer
+                            ctx={this.ctx}
+                            node={this.node}
+                            view={this.view}
+                            getPos={this.getPos}
+                            decorations={this.decorations}
+                        >
+                            <CustomComponent>
+                                <Content isMark={this.node instanceof Mark} />
+                            </CustomComponent>
+                        </VueNodeContainer>
+                    </Teleport>
+                );
+            },
         });
-        this.addPortal(Portal, this.key);
+        this.addPortal(markRaw(Portal) as DefineComponent, this.key);
+        const instance = getRootInstance();
+        if (instance) {
+            instance.update();
+        }
     }
 
     destroy() {
-        this.dom = undefined;
-        this.contentDOM = undefined;
         this.removePortalByKey(this.key);
     }
 
     ignoreMutation(mutation: MutationRecord | { type: 'selection'; target: Element }) {
-        if (!this.contentDOM) {
+        if (!this.dom || !this.contentDOM) {
             return true;
         }
-        return !this.contentDOM.contains(mutation.target);
+
+        if (this.node instanceof Node) {
+            if (this.node.isLeaf || this.node.isAtom) {
+                return true;
+            }
+        }
+
+        if (mutation.type === 'selection') {
+            return false;
+        }
+
+        if (this.contentDOM === this.dom) {
+            return false;
+        }
+
+        if (this.contentDOM.contains(mutation.target)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    update(node: Node | Mark, decorations: Decoration[]) {
+        if (this.node.type !== node.type) {
+            return false;
+        }
+
+        if (node === this.node && this.decorations === decorations) {
+            return true;
+        }
+
+        this.node = node;
+        this.decorations = decorations;
+        return true;
     }
 }