+++ packages/vue/vite.config.ts
@@ -2,6 +2,8 @@
 import vueJsx from '@vitejs/plugin-vue-jsx';
 import { defineConfig } from 'vite';
 
+import { viteBuild } from '../../vite.config.common';
+
 export default defineConfig({
     root: 'app',
     plugins: [vueJsx()],
@@ -9,4 +11,5 @@ export default defineConfig({
         jsxFactory: 'h',
         jsxFragment: 'Fragment',
     },
+    build: viteBuild('vue'),
 });