+++ packages/preset-gfm/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @milkdown/preset-gfm
 
+## 5.3.0
+
+### Minor Changes
+
+-   ed679a03: Optimize vue and react renderer.
+
+### Patch Changes
+
+-   Updated dependencies [ed679a03]
+    -   @milkdown/plugin-table@5.3.0
+    -   @milkdown/preset-commonmark@5.3.0
+    -   @milkdown/utils@5.3.0
+
+## 5.2.1
+
+### Patch Changes
+
+-   eb74e3b5: Add es and cjs bundle, fix bugs and improve ux
+-   Updated dependencies [eb74e3b5]
+    -   @milkdown/plugin-table@5.2.1
+    -   @milkdown/preset-commonmark@5.2.1
+    -   @milkdown/utils@5.2.1
+
+## 5.2.0
+
+### Minor Changes
+
+-   42055660: Add menu plugin and lots of optimize.
+
+### Patch Changes
+
+-   Updated dependencies [42055660]
+    -   @milkdown/plugin-table@5.2.0
+    -   @milkdown/preset-commonmark@5.2.0
+    -   @milkdown/utils@5.2.0
+
 ## 5.1.2
 
 ### Patch Changes